Move Twig classes to PhpMyAdmin namespace
