+++ scripts/generate-twig-cache
@@ -4,12 +4,12 @@ define('PHPMYADMIN', 1);
 require_once 'libraries/vendor_config.php';
 require_once AUTOLOAD_FILE;
 
-use PMA\libraries\twig\CharsetsExtension;
-use PMA\libraries\twig\I18nExtension;
-use PMA\libraries\twig\MessageExtension;
-use PMA\libraries\twig\SanitizeExtension;
-use PMA\libraries\twig\UrlExtension;
-use PMA\libraries\twig\UtilExtension;
+use PhpMyAdmin\Twig\CharsetsExtension;
+use PhpMyAdmin\Twig\I18nExtension;
+use PhpMyAdmin\Twig\MessageExtension;
+use PhpMyAdmin\Twig\SanitizeExtension;
+use PhpMyAdmin\Twig\UrlExtension;
+use PhpMyAdmin\Twig\UtilExtension;
 
 $tplDir = dirname(__FILE__) . '/../templates';
 $tmpDir = dirname(__FILE__) . '/../twig-templates';
